# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BeerOut": {
        "additionalProperties": false,
        "properties": {
          "abv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abv"
          },
          "average_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Rating"
          },
          "brewery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BreweryRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "brewery_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brewery Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "style": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StyleSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "tastings_count": {
            "title": "Tastings Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "brewery_name",
          "style",
          "abv",
          "notes",
          "brewery",
          "tastings_count",
          "average_rating",
          "created_at",
          "updated_at"
        ],
        "title": "BeerOut",
        "type": "object"
      },
      "BeerUpdate": {
        "additionalProperties": false,
        "properties": {
          "abv": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abv"
          },
          "brewery_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brewery Id"
          },
          "brewery_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brewery Name"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "style": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          }
        },
        "title": "BeerUpdate",
        "type": "object"
      },
      "BeerWrite": {
        "additionalProperties": false,
        "properties": {
          "abv": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abv"
          },
          "brewery_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A brewery from the map",
            "title": "Brewery Id"
          },
          "brewery_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brewery Name"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "maxLength": 10000,
            "title": "Notes",
            "type": "string"
          },
          "style": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Style slug",
            "title": "Style"
          }
        },
        "required": [
          "name"
        ],
        "title": "BeerWrite",
        "type": "object"
      },
...
      "BreweryRef": {
        "additionalProperties": false,
        "description": "Enough to name a brewery next to a beer or tasting.",
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state_province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Province"
          }
        },
        "required": [
          "id",
          "name",
          "city",
          "state_province",
          "country"
        ],
        "title": "BreweryRef",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "loc": {
            "description": "Path to the offending field",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Page_BeerOut_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BeerOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as ?cursor= to fetch the next page; null on the last page",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[BeerOut]",
        "type": "object"
      },
...
      "Problem": {
        "description": "Problem details object (RFC 9457) with a request_id extension member.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "StyleRangeOut": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "max": {
            "title": "Max",
            "type": "number"
          },
          "metric": {
            "enum": [
              "og",
              "fg",
              "abv",
              "ibu",
              "srm"
            ],
            "title": "Metric",
            "type": "string"
          },
          "min": {
            "title": "Min",
            "type": "number"
          }
        },
        "required": [
          "metric",
          "min",
          "max",
          "label"
        ],
        "title": "StyleRangeOut",
        "type": "object"
      },
      "StyleSummary": {
        "additionalProperties": false,
        "properties": {
          "category_code": {
            "title": "Category Code",
            "type": "string"
          },
          "category_name": {
            "title": "Category Name",
            "type": "string"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Slug"
          },
          "ranges": {
            "items": {
              "$ref": "#/components/schemas/StyleRangeOut"
            },
            "title": "Ranges",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "code",
          "name",
          "display_name",
          "category_code",
          "category_name",
          "parent_slug",
          "ranges"
        ],
        "title": "StyleSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Recipe design, brewing log, tasting notes and beer discovery.",
    "title": "BrewNotes API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/beers": {
      "get": {
        "operationId": "list_beers_api_v1_beers_get",
        "parameters": [
          {
            "description": "Match on the beer or brewery name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Match on the beer or brewery name",
              "title": "Q"
            }
          },
          {
            "description": "Only tastings of this subject",
            "in": "query",
            "name": "brewery_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tastings of this subject",
              "title": "Brewery Id"
            }
          },
          {
            "description": "Opaque cursor from a previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 400,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BeerOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Client error (problem details)"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error (problem details)"
          }
        },
        "summary": "List my beers alphabetically",
        "tags": [
          "beers"
        ]
      },
      "post": {
        "operationId": "create_beer_api_v1_beers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Client error (problem details)"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error (problem details)"
          }
        },
        "summary": "Add a commercial beer",
        "tags": [
          "beers"
        ]
      }
    },
    "/api/v1/beers/{beer_id}": {
      "delete": {
        "operationId": "delete_beer_api_v1_beers__beer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "beer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Beer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Client error (problem details)"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error (problem details)"
          }
        },
        "summary": "Delete a beer and its tastings",
        "tags": [
          "beers"
        ]
      },
      "get": {
        "operationId": "read_beer_api_v1_beers__beer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "beer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Beer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Client error (problem details)"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error (problem details)"
          }
        },
        "summary": "A beer with its tasting count and average rating",
        "tags": [
          "beers"
        ]
      },
      "patch": {
        "operationId": "update_beer_api_v1_beers__beer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "beer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Beer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Client error (problem details)"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error (problem details)"
          }
        },
        "summary": "Update a beer",
        "tags": [
          "beers"
        ]
      }
    },
...
  }
}